+++ package.json
@@ -44,6 +44,8 @@
     "webpack-dev-server": "^3.11.2"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.12.1"
+    "@babel/polyfill": "^7.12.1",
+    "jquery": "^3.5.1",
+    "slick-carousel": "^1.8.1"
   }
 }
