+++ src/scss/components/_header.scss
@@ -48,7 +48,7 @@
                 &::before {
                     display: block;
                 }
-             }
+            }
         }
     }
